requirements.txt
aiohttp==3.9.1
pyquery==2.0.0 
discord.py==2.3.2
python-dotenv==1.0.0
selenium==4.18.1
webdriver-manager==4.0.1 